+++ src/ffi/mod.rs
@@ -3,6 +3,7 @@
 mod advanced;
 mod command;
 mod core;
+mod exhaustive;
 mod metalfx;
 mod pipeline;
 mod render;
@@ -11,6 +12,7 @@ mod state;
 pub use advanced::*;
 pub use command::*;
 pub use core::*;
+pub use exhaustive::*;
 pub use metalfx::*;
 pub use pipeline::*;
 pub use render::*;